Extract exception classes to server/exceptions.js
